On all bindings manual pages I could find (example: http://gridscheduler.sourceforge.net/htmlman/htmlman3/drmaa_attributes.html) the drmaa_get_num_xxx functions uses a int * to store their result. On some (most?) platforms, int * and size_t are not compatibles. The go binding, for one, fails to build because of this.

+++ drmaa_utils/drmaa_base.c
@@ -277,7 +277,7 @@ int drmaa_get_next_##name( type *values, char *value, size_t value_len ) \
 	strlcpy( value, iter->next(iter), value_len ); \
 	DRMAA_API_END \
 } \
-int drmaa_get_num_##name##s( type *values, size_t *size ) \
+int drmaa_get_num_##name##s( type *values, int *size ) \
 { \
 	char error_diagnosis[1]; \
 	size_t error_diag_len = sizeof(error_diagnosis); \
